+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- YamlLoader to fail fast or return default if specified.
+
 ## [0.1.5] - 2024-06-09
 
 ### Added